Make AgentConnect integration example more accessible
The AgentConnect button example in the docs missed alternative text for the "S'identifier avec AgentConnect" css background image. This meant that if I can't see the image (because I disabled CSS, or I use a screen reader like VoiceOver), I couldn't get what was the button about.

Now we use a classic "screen reader only" class to write the text in the DOM without rendering it.